Check which points are located on land.

Check which points are located on land.

check_onland(data, land = NULL, report = FALSE, buffer = 0,
  offline = FALSE)

Arguments

data
The data frame.

land
SpatialPolygonsDataFrame. If not provided the simplified land polygons from OSM are used. This parameter is ignored when, offline = FALSE.

report
If TRUE, errors are returned instead of records.

buffer
Set how far inland points are still to be deemed valid (in meters).

offline
If TRUE, a local simplified shoreline is used, otherwise an OBIS webservice is used. The default value is FALSE.

Value

Errors or problematic records.

See also

check_depth check_outliers_dataset check_outliers_species lookup_xy

Examples

# NOT RUN {
report <- check_onland(abra, report = TRUE, buffer = 100)
print(report)
# plot records on land with 100 meter buffer
plot_map_leaflet(abra[report$row,], popup = "id")
# filter records not on land
ok <- abra[-1 * report$row,]
ok <- check_onland(abra, report = FALSE, buffer = 100)
print(nrow(ok))
# }