Lookup spatial data for a set of points.

Lookup spatial data for a set of points.

lookup_xy(data, shoredistance=TRUE, grids=TRUE, areas=FALSE,
  asdataframe=TRUE)

Arguments

data
The data frame with columns decimalLongitude and decimalLatitude.

shoredistance
Indicate whether the shoredistance should be returned (default TRUE).

grids
Indictate whether the grid values such as temperature and bathymetry should be returned (default TRUE)

areas
Indictate whether the area values should be returned (default FALSE).

asdataframe
Indicate whether a dataframe or a list should be returned (default TRUE).

Value

Data frame or list with the values for the different requested fields.

Details

When asdataframe is FALSE then data is returned in the same order as the requested data as a list, with for each list item the requested values. For invalid coordinates NULL is returned.

See also

check_onland check_depth

Examples

lookup_xy(abra, shoredistance = TRUE, grids = TRUE, areas = FALSE)

#>      shoredistance sstemperature sssalinity    bathymetry
#> 1               30     10.286312  34.762711 -4.000000e+00
#> 2             1080     10.332419  34.906219  6.140000e+01
#> 3             1184     10.721993  34.888958  1.222000e+02
#> 4              290     10.791965  34.293419  2.060000e+01
#> 5              259     10.721993  34.888958  5.100000e+01
#> 6              506     10.771008  34.306995  3.240000e+01
#> 7              616     10.771008  34.306995  2.880000e+01
#> 8               30     10.286312  34.762711 -4.000000e+00
#> 9             1489     10.332730  34.858673  6.600000e+01
#> 10             417     10.771008  34.306995  2.240000e+01
#> 11            1385     10.721993  34.888958  1.792000e+02
#> 12            5634     10.385038  34.864372  1.114000e+02
#> 13             565     10.566178  34.661377  3.480000e+01
#> 14             723     10.771008  34.306995  2.180000e+01
#> 15            1018     10.332419  34.906219  8.600000e+01
#> 16             331     10.530102  34.685471  7.800000e+00
#> 17             468     10.721993  34.888958  9.240000e+01
#> 18             175            NA         NA  4.400000e+00
#> 19             292     10.302540  34.743073  1.420000e+01
#> 20             351     10.771008  34.306995  2.480000e+01
#> 21             292     10.302540  34.743073  1.420000e+01
#> 22             587     10.791965  34.293419  2.600000e+01
#> 23             916     10.771008  34.306995  2.640000e+01
#> 24              98     10.479015  34.686848 -1.500000e+01
#> 25             600     10.286147  34.834846  4.200000e+00
#> 26             261     10.721993  34.888958  5.540000e+01
#> 27             370     10.519511  34.687080  2.400000e+00
#> 28             752     10.791965  34.293419  2.380000e+01
#> 29             525     10.329433  34.895565 -1.420000e+01
#> 30            6371     10.343019  34.140945  4.660000e+01
#> 31            -194     10.400653  34.231487 -6.600000e+01
#> 32            3146     12.482795  33.972263  6.400000e+00
#> 33            2811     10.163292  18.081486  1.460000e+01
#> 34            2268     10.332844  15.748602  1.460000e+01
#> 35            2268     10.332844  15.748602  1.460000e+01
#> 36            3577     10.231925  20.753168  2.020000e+01
#> 37            3577     10.231925  20.753168  2.020000e+01
#> 38            2811     10.163292  18.081486  1.460000e+01
#> 39            2811     10.163292  18.081486  1.460000e+01
#> 40            2811     10.163292  18.081486  1.460000e+01
#> 41            2811     10.163292  18.081486  1.460000e+01
#> 42            2811     10.163292  18.081486  1.460000e+01
#> 43            2811     10.163292  18.081486  1.460000e+01
#> 44            2811     10.163292  18.081486  1.460000e+01
#> 45            2811     10.163292  18.081486  1.460000e+01
#> 46            3884     10.016996  16.595095  1.860000e+01
#> 47            3884     10.016996  16.595095  1.860000e+01
#> 48            4630     10.031013  17.248009  1.640000e+01
#> 49            4630     10.031013  17.248009  1.640000e+01
#> 50            4630     10.031013  17.248009  1.640000e+01
#> 51            3884     10.016996  16.595095  1.860000e+01
#> 52            3884     10.016996  16.595095  1.860000e+01
#> 53           13666     10.234084  22.443727  1.300000e+01
#> 54             692     10.022824  17.227293  7.200000e+00
#> 55             692     10.022824  17.227293  7.200000e+00
#> 56             692     10.022824  17.227293  7.200000e+00
#> 57             692     10.022824  17.227293  7.200000e+00
#> 58             504     10.210454  15.189676  4.600000e+00
#> 59             696     10.415621  12.795743  1.180000e+01
#> 60            3577     10.231925  20.753168  2.020000e+01
#> 61            3689     10.384777  16.132429  1.220000e+01
#> 62            3689     10.384777  16.132429  1.220000e+01
#> 63            2606     10.251281  18.826040  1.400000e+01
#> 64            2606     10.251281  18.826040  1.400000e+01
#> 65            2606     10.251281  18.826040  1.400000e+01
#> 66            2606     10.251281  18.826040  1.400000e+01
#> 67            2606     10.251281  18.826040  1.400000e+01
#> 68            2606     10.251281  18.826040  1.400000e+01
#> 69            2606     10.251281  18.826040  1.400000e+01
#> 70            2606     10.251281  18.826040  1.400000e+01
#> 71            2606     10.251281  18.826040  1.400000e+01
#> 72            2606     10.251281  18.826040  1.400000e+01
#> 73            1326     10.340107  12.981791  7.400000e+00
#> 74            1326     10.340107  12.981791  7.400000e+00
#> 75            1326     10.340107  12.981791  7.400000e+00
#> 76            2606     10.251281  18.826040  1.400000e+01
#> 77            2606     10.251281  18.826040  1.400000e+01
#> 78            2606     10.251281  18.826040  1.400000e+01
#> 79            2606     10.251281  18.826040  1.400000e+01
#> 80            2606     10.251281  18.826040  1.400000e+01
#> 81            2606     10.251281  18.826040  1.400000e+01
#> 82            4934     10.077607  17.306303  1.900000e+01
#> 83            4934     10.077607  17.306303  1.900000e+01
#> 84            4934     10.077607  17.306303  1.900000e+01
#> 85            4934     10.077607  17.306303  1.900000e+01
#> 86            4934     10.077607  17.306303  1.900000e+01
#> 87            4934     10.077607  17.306303  1.900000e+01
#> 88            4934     10.077607  17.306303  1.900000e+01
#> 89            4934     10.077607  17.306303  1.900000e+01
#> 90            4934     10.077607  17.306303  1.900000e+01
#> 91            4934     10.077607  17.306303  1.900000e+01
#> 92            4934     10.077607  17.306303  1.900000e+01
#> 93            4934     10.077607  17.306303  1.900000e+01
#> 94            4934     10.077607  17.306303  1.900000e+01
#> 95            4934     10.077607  17.306303  1.900000e+01
#> 96            4934     10.077607  17.306303  1.900000e+01
#> 97            4934     10.077607  17.306303  1.900000e+01
#> 98            4934     10.077607  17.306303  1.900000e+01
#> 99            4934     10.077607  17.306303  1.900000e+01
#> 100           4934     10.077607  17.306303  1.900000e+01
#> 101           4934     10.077607  17.306303  1.900000e+01
#> 102           4934     10.077607  17.306303  1.900000e+01
#> 103           4934     10.077607  17.306303  1.900000e+01
#> 104           4934     10.077607  17.306303  1.900000e+01
#> 105           4934     10.077607  17.306303  1.900000e+01
#> 106           4934     10.077607  17.306303  1.900000e+01
#> 107           4934     10.077607  17.306303  1.900000e+01
#> 108           4934     10.077607  17.306303  1.900000e+01
#> 109           4934     10.077607  17.306303  1.900000e+01
#> 110           4934     10.077607  17.306303  1.900000e+01
#> 111           4934     10.077607  17.306303  1.900000e+01
#> 112           4934     10.077607  17.306303  1.900000e+01
#> 113           4934     10.077607  17.306303  1.900000e+01
#> 114           4934     10.077607  17.306303  1.900000e+01
#> 115           4934     10.077607  17.306303  1.900000e+01
#> 116           2606     10.251281  18.826040  1.400000e+01
#> 117           2606     10.251281  18.826040  1.400000e+01
#> 118            889     10.414397  12.669686  1.040000e+01
#> 119            889     10.414397  12.669686  1.040000e+01
#> 120            889     10.414397  12.669686  1.040000e+01
#> 121            889     10.414397  12.669686  1.040000e+01
#> 122            889     10.414397  12.669686  1.040000e+01
#> 123            889     10.414397  12.669686  1.040000e+01
#> 124           4873     10.257658  12.240781  1.220000e+01
#> 125           5920     10.276494  14.850209  1.900000e+01
#> 126           5920     10.276494  14.850209  1.900000e+01
#> 127           5920     10.276494  14.850209  1.900000e+01
#> 128           5920     10.276494  14.850209  1.900000e+01
#> 129           5920     10.276494  14.850209  1.900000e+01
#> 130           5920     10.276494  14.850209  1.900000e+01
#> 131            570            NA         NA  5.000000e+00
#> 132           5920     10.276494  14.850209  1.900000e+01
#> 133           5920     10.276494  14.850209  1.900000e+01
#> 134           5920     10.276494  14.850209  1.900000e+01
#> 135           5920     10.276494  14.850209  1.900000e+01
#> 136           5920     10.276494  14.850209  1.900000e+01
#> 137           5920     10.276494  14.850209  1.900000e+01
#> 138           5920     10.276494  14.850209  1.900000e+01
#> 139           5920     10.276494  14.850209  1.900000e+01
#> 140           5920     10.276494  14.850209  1.900000e+01
#> 141           5920     10.276494  14.850209  1.900000e+01
#> 142            570            NA         NA  5.000000e+00
#> 143           3577     10.231925  20.753168  2.020000e+01
#> 144           4150     10.387556  20.308191  1.700000e+01
#> 145           4150     10.387556  20.308191  1.700000e+01
#> 146           4150     10.387556  20.308191  1.700000e+01
#> 147           4150     10.387556  20.308191  1.700000e+01
#> 148           4150     10.387556  20.308191  1.700000e+01
#> 149           4150     10.387556  20.308191  1.700000e+01
#> 150           4150     10.387556  20.308191  1.700000e+01
#> 151           6937     10.251081  13.376673  2.320000e+01
#> 152           6937     10.251081  13.376673  2.320000e+01
#> 153           6937     10.251081  13.376673  2.320000e+01
#> 154           6937     10.251081  13.376673  2.320000e+01
#> 155           6937     10.251081  13.376673  2.320000e+01
#> 156           6937     10.251081  13.376673  2.320000e+01
#> 157           6937     10.251081  13.376673  2.320000e+01
#> 158           4150     10.387556  20.308191  1.700000e+01
#> 159            504     10.210454  15.189676  4.600000e+00
#> 160            504     10.210454  15.189676  4.600000e+00
#> 161            504     10.210454  15.189676  4.600000e+00
#> 162            504     10.210454  15.189676  4.600000e+00
#> 163            504     10.210454  15.189676  4.600000e+00
#> 164            504     10.210454  15.189676  4.600000e+00
#> 165           3689     10.384777  16.132429  1.220000e+01
#> 166           3689     10.384777  16.132429  1.220000e+01
#> 167           3689     10.384777  16.132429  1.220000e+01
#> 168           3689     10.384777  16.132429  1.220000e+01
#> 169           3689     10.384777  16.132429  1.220000e+01
#> 170           3689     10.384777  16.132429  1.220000e+01
#> 171            692     10.022824  17.227293  7.200000e+00
#> 172            570            NA         NA  5.000000e+00
#> 173            570            NA         NA  5.000000e+00
#> 174            570            NA         NA  5.000000e+00
#> 175            570            NA         NA  5.000000e+00
#> 176            570            NA         NA  5.000000e+00
#> 177           2606     10.251281  18.826040  1.400000e+01
#> 178           3051      9.983971  17.246365  1.800000e+01
#> 179           3051      9.983971  17.246365  1.800000e+01
#> 180           3051      9.983971  17.246365  1.800000e+01
#> 181           2606     10.251281  18.826040  1.400000e+01
#> 182           2606     10.251281  18.826040  1.400000e+01
#> 183           2606     10.251281  18.826040  1.400000e+01
#> 184           2606     10.251281  18.826040  1.400000e+01
#> 185           2606     10.251281  18.826040  1.400000e+01
#> 186           2606     10.251281  18.826040  1.400000e+01
#> 187           2606     10.251281  18.826040  1.400000e+01
#> 188           2606     10.251281  18.826040  1.400000e+01
#> 189           2606     10.251281  18.826040  1.400000e+01
#> 190           2606     10.251281  18.826040  1.400000e+01
#> 191           2606     10.251281  18.826040  1.400000e+01
#> 192           2606     10.251281  18.826040  1.400000e+01
#> 193           2606     10.251281  18.826040  1.400000e+01
#> 194           2606     10.251281  18.826040  1.400000e+01
#> 195           2606     10.251281  18.826040  1.400000e+01
#> 196           2606     10.251281  18.826040  1.400000e+01
#> 197           2268     10.332844  15.748602  1.460000e+01
#> 198           2268     10.332844  15.748602  1.460000e+01
#> 199           2606     10.251281  18.826040  1.400000e+01
#> 200           2606     10.251281  18.826040  1.400000e+01
#> 201           2606     10.251281  18.826040  1.400000e+01
#> 202           2606     10.251281  18.826040  1.400000e+01
#> 203           2606     10.251281  18.826040  1.400000e+01
#> 204           2606     10.251281  18.826040  1.400000e+01
#> 205           2606     10.251281  18.826040  1.400000e+01
#> 206           2606     10.251281  18.826040  1.400000e+01
#> 207           2606     10.251281  18.826040  1.400000e+01
#> 208           2606     10.251281  18.826040  1.400000e+01
#> 209           2606     10.251281  18.826040  1.400000e+01
#> 210           2606     10.251281  18.826040  1.400000e+01
#> 211           2606     10.251281  18.826040  1.400000e+01
#> 212           2606     10.251281  18.826040  1.400000e+01
#> 213           2606     10.251281  18.826040  1.400000e+01
#> 214           2606     10.251281  18.826040  1.400000e+01
#> 215           2606     10.251281  18.826040  1.400000e+01
#> 216           2606     10.251281  18.826040  1.400000e+01
#> 217           2606     10.251281  18.826040  1.400000e+01
#> 218           2606     10.251281  18.826040  1.400000e+01
#> 219           2606     10.251281  18.826040  1.400000e+01
#> 220           1266     10.521104  33.046227 -4.000000e-01
#> 221            570            NA         NA  5.000000e+00
#> 222            504     10.210454  15.189676  4.600000e+00
#> 223           7223     10.396317  20.349792  1.460000e+01
#> 224           7223     10.396317  20.349792  1.460000e+01
#> 225           3051      9.983971  17.246365  1.800000e+01
#> 226           3051      9.983971  17.246365  1.800000e+01
#> 227           3051      9.983971  17.246365  1.800000e+01
#> 228           3051      9.983971  17.246365  1.800000e+01
#> 229           3051      9.983971  17.246365  1.800000e+01
#> 230           3051      9.983971  17.246365  1.800000e+01
#> 231           3051      9.983971  17.246365  1.800000e+01
#> 232           3051      9.983971  17.246365  1.800000e+01
#> 233           3051      9.983971  17.246365  1.800000e+01
#> 234           3051      9.983971  17.246365  1.800000e+01
#> 235           3689     10.384777  16.132429  1.220000e+01
#> 236           3689     10.384777  16.132429  1.220000e+01
#> 237           3689     10.384777  16.132429  1.220000e+01
#> 238            692     10.022824  17.227293  7.200000e+00
#> 239            692     10.022824  17.227293  7.200000e+00
#> 240           3051      9.983971  17.246365  1.800000e+01
#> 241           3051      9.983971  17.246365  1.800000e+01
#> 242           3051      9.983971  17.246365  1.800000e+01
#> 243           3051      9.983971  17.246365  1.800000e+01
#> 244           3051      9.983971  17.246365  1.800000e+01
#> 245           4630     10.031013  17.248009  1.640000e+01
#> 246           4630     10.031013  17.248009  1.640000e+01
#> 247           4630     10.031013  17.248009  1.640000e+01
#> 248           4630     10.031013  17.248009  1.640000e+01
#> 249           4630     10.031013  17.248009  1.640000e+01
#> 250            504     10.210454  15.189676  4.600000e+00
#>  [ reached getOption("max.print") -- omitted 1843 rows ]