Lookup spatial data for a set of points.

Lookup spatial data for a set of points.

lookup_xy(data, shoredistance=TRUE, grids=TRUE, areas=FALSE,
  asdataframe=TRUE)

Arguments

data
The data frame with columns decimalLongitude and decimalLatitude.

shoredistance
Indicate whether the shoredistance should be returned (default TRUE).

grids
Indicate whether the grid values such as temperature and bathymetry should be returned (default TRUE)

areas
Indicate whether the area values should be returned and from which distance in meters to the provided points (default FALSE).

asdataframe
Indicate whether a dataframe or a list should be returned (default TRUE).

Value

Data frame or list with the values for the different requested fields.

Details

When asdataframe is FALSE then data is returned in the same order as the requested data as a list, with for each list item the requested values. For invalid coordinates NULL is returned.

When parameter areas is a positive integer then all areas within a radius of this distance in meters will be returned. A value of TRUE is equivalent to a distance of 0 meters, FALSE indicates that no area results are required.

See also

check_onland check_depth

Examples

# NOT RUN {
lookup_xy(abra, shoredistance = TRUE, grids = TRUE, areas = FALSE)
# }